wireless: fix rtw_hal_com_set_gt3 timeout type (u8 -> u32)
rtw_hal_com_set_gt3() defined its timeout argument as u8 while every
other point in the call chain uses u32:

  caller _phl_set_gt3_with_type()  : u32 timeout  (core/.../phl_tx.c)
  prototype in hal_api.h           : u32 timeout
  callee rtw_hal_mac_set_gt3()     : u32 timeout

The u8 definition silently truncated the value before handing it to the
MAC layer, so any GT3 timeout above 255 was clipped. The prototype and
the rest of the chain are correct; the definition was wrong. Widen it to
u32 to match.

This mismatch was invisible until the prototype became visible at the
definition site (C only type-checks within a translation unit, and the
linker matches by symbol name only).

+++ phl/hal_g6/hal_com_i.c
@@ -85,7 +85,7 @@ rtw_hal_com_set_power_offset(void *hal, u8 band, s8 ofst_mode, s8 ofst_bw)
 	return RTW_HAL_STATUS_SUCCESS;
 }
 
-enum rtw_hal_status rtw_hal_com_set_gt3(void *hal, u8 en, u8 timeout)
+enum rtw_hal_status rtw_hal_com_set_gt3(void *hal, u8 en, u32 timeout)
 {
 	struct hal_info_t *hal_info = (struct hal_info_t *)hal;
